Move KeyUtil from util module to test module
